requirements.txt
channels
channels-redis
uvicorn
Django<3.0
django-admin-sortable2
django-cors-headers
django-mysql
djangorestframework
mysqlclient
redis
rq
